backend/requirements.txt
chromadb 
google-genai 
networkx 
python-dotenv
fastapi
uvicorn